i386: Fix vect-pragma-target-[12].c testcase for -march=XYZ [PR120643]
These 2 testcases were originally designed for the default -march= of
x86_64 so if you pass -march=native (on a target with AVX512 enabled),
they will fail. It fix this, we add `-mno-sse3 -mtune=generic`
to the options to force a specific arch to the testcase.

Changes since v1:
* v2: Use -mtune=generic instead of -mprefer-vector-width=512.

Tested on a skylake-avx512 machine with -march=native.

	PR testsuite/120643
gcc/testsuite/ChangeLog:

	* gcc.target/i386/vect-pragma-target-1.c: Add `-mno-sse3 -mtune=generic`
	to the options.
	* gcc.target/i386/vect-pragma-target-2.c: Likewise.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>