configure, Darwin: Do not claim .cfi_xxx instruction support.
While the assemblers used by Darwin that are based on LLVM, do
support .cfi_ instructions, their use triggers production of
compact unwind which currently does not interwork properly with
GCC's output.

When the system objdump is used in the configure process this is
currently working by good fortune (the objdump does not recognise
the command and we fail to detect the cfi_advance.

However, if a user has binutils objdump earlier in thier PATH then
we will detect support and try to use .cfi_ which will cause later
and hard-to-diagnose issues.

Until we have this resolved, force cfi instruction use off for
Darwin.

gcc/ChangeLog:

	* configure: Regenerate.
	* configure.ac: Do not claim cfi instruction support even
	if the assembler has it.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>