diagnostics: Fix bootstrap fail on Darwin 32b hosts.
The use of HOST_SIZE_T_PRINT_HEX needs to be paired with a c-style
cast to (fmt_size_t) otherwise the detection mechanisms in hwint.h
are not sufficient to deal with size_t defined as 'long unsigned int'
which is done on Darwin (and I think on Windows).

This patch just makes that update.

gcc/ChangeLog:

	* diagnostics/logging.h (log_param_location_t): Cast
	location_t value to fmt_size_t.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>