s390: Adjust s390/spaceship-fp-*.c tests for recent changes
In r16-3414 libstdc++ changed ABI for (still experimental C++20) and uses
unordered value -128 instead of 2.  Generally the change improved code
generation on all targets tested, see
https://gcc.gnu.org/pipermail/gcc-patches/2025-August/693534.html
for details.
In r16-3474 I've adjusted the middle-end and backends to use that value.
This apparently broke the gcc.target/s390/spaceship-fp-2.c test,
with -ffast-math the 2 value is unreachable and so the .SPACESHIP last
argument in that case is the default, which changed from 2 to -128.
But spaceship-fp-1.c test also doesn't test what libstdc++ uses anymore,
so the following patch uses -128 in all the spots.

2025-09-02  Jakub Jelinek  <jakub@redhat.com>

	* gcc.target/s390/spaceship-fp-1.c: Expect .SPACESHIP call with
	-128 as last argument instead of 2.
	(TEST): Use -128 instead of 2.
	* gcc.target/s390/spaceship-fp-2.c: Expect .SPACESHIP call with
	-128 as last argument instead of 2.
	(TEST): Use -128 instead of 2.