libstdc++: Restore C++20 <chrono> support for old std::string ABI
The r16-3416-g806de30f51c8b9 change to use __cpp_lib_chrono in
preprocessor conditions broke support for <chrono> for freestanding and
the COW std::string ABI. That happened because __cpp_lib_chrono is only
defined to the C++20 value for hosted and for the new ABI, because the
full set of C++20 features are not defined for freestanding and tzdb is
not defined for the old ABI.

This introduces a new internal feature test macro that corresponds to
the features that are always supported (e.g. chrono::local_time,
chrono::year, chrono::weekday).

libstdc++-v3/ChangeLog:

	* include/bits/version.def (chrono_cxx20): Define.
	* include/bits/version.h: Regenerate.
	* include/std/chrono: Check __glibcxx_chrono_cxx20 instead of
	__cpp_lib_chrono for C++20 features that don't require the new
	std::string ABI and/or can be used for freestanding.
	* src/c++20/clock.cc: Adjust preprocessor condition.

Reviewed-by: Tomasz Kamiński <tkaminsk@redhat.com>