Increase default number of LTO partitions
The number of LTO partitions should exceed number of CPUs (or hyper-threads) of
commonly used CPUs.  I think it is time to increase it again and as discussed
in the LTO and toplevel asm thread, doing so scales quite well. Tmp file usage
grows from 2.7 to 2.9MB which seems acceptable.  Overall build time on machine
with 256 hyperthreads is comparable.

Bootstrapped/regtested x86_64-linux, comitted.

gcc/ChangeLog:

	* params.opt (-param=lto-partitions=): INcrease default value from 128 to 512.