arm: wrong code from vset_lane_* [PR121775]
Insufficient validation of the operands in vec_set_<mode>_internal
means that the optimizers can transform the exanded code into
something that is invalid.  We then emit code based on the incorrect
RTL assuming that it is still valid.  A valid pattern can only have a
single bit set in the immediate operand, representing the lane to be
written.

gcc/ChangeLog:

	PR target/121775
	* config/arm/neon.md (vec_set<mode>_internal, all variants):
	validate the immediate operand that indicates the lane to
	modify.

gcc/testsuite/ChangeLog:

	PR target/121775
	* gcc.target/arm/simd/vset_lane_u8.c: New test.